Revert "MINOR: Improve UDF Lineage Processing & Better Logging Time & MultiProcessing (#20848)"
This reverts commit 5ea9f22492749867f9ea53465b817f52bd383ca2.
