+++ ingestion/src/metadata/ingestion/lineage/masker.py
@@ -20,7 +20,6 @@ from sqlparse.sql import Comparison
 from sqlparse.tokens import Literal, Number, String
 
 from metadata.ingestion.lineage.models import Dialect
-from metadata.utils.execution_time_tracker import calculate_execution_time
 
 MASK_TOKEN = "?"
 
@@ -113,12 +112,8 @@ def mask_literals_with_sqlfluff(query: str, parser: LineageRunner) -> str:
     return query
 
 
-@calculate_execution_time(context="MaskQuery")
 def mask_query(
-    query: str,
-    dialect: str = Dialect.ANSI.value,
-    parser: LineageRunner = None,
-    parser_required: bool = False,
+    query: str, dialect: str = Dialect.ANSI.value, parser: LineageRunner = None
 ) -> str:
     """
     Mask a query using sqlparse or sqlfluff.
@@ -127,8 +122,6 @@ def mask_query(
     try:
         if masked_query_cache.get((query, dialect)):
             return masked_query_cache.get((query, dialect))
-        if parser_required and not parser:
-            return None
         if not parser:
             try:
                 parser = LineageRunner(query, dialect=dialect)
