+++ ingestion/src/metadata/ingestion/source/database/lineage_source.py
@@ -12,21 +12,22 @@
 Lineage Source Module
 """
 import csv
-import multiprocessing
 import os
+import threading
 import time
 import traceback
 from abc import ABC
 from functools import partial
-from multiprocessing import Process, Queue
-from typing import Any, Callable, Iterable, Iterator, List, Optional, Tuple, Union
-
-import networkx as nx
+from typing import Any, Callable, Iterable, Iterator, List, Optional, Union
 
 from metadata.generated.schema.api.data.createQuery import CreateQueryRequest
 from metadata.generated.schema.api.lineage.addLineage import AddLineageRequest
 from metadata.generated.schema.entity.data.table import Table
-from metadata.generated.schema.type.basic import Uuid
+from metadata.generated.schema.type.basic import (
+    FullyQualifiedEntityName,
+    SqlQuery,
+    Uuid,
+)
 from metadata.generated.schema.type.entityLineage import (
     ColumnLineage,
     EntitiesEdge,
@@ -37,24 +38,25 @@ from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.generated.schema.type.tableQuery import TableQuery
 from metadata.ingestion.api.models import Either
 from metadata.ingestion.lineage.models import ConnectionTypeDialectMapper, Dialect
-from metadata.ingestion.lineage.sql_lineage import get_column_fqn, get_lineage_by_graph
-from metadata.ingestion.source.database.lineage_processors import (
-    _process_chunk_in_subprocess,
-    query_lineage_generator,
-    view_lineage_generator,
+from metadata.ingestion.lineage.sql_lineage import (
+    get_column_fqn,
+    get_lineage_by_graph,
+    get_lineage_by_query,
 )
+from metadata.ingestion.models.ometa_lineage import OMetaLineageRequest
+from metadata.ingestion.models.topology import Queue
 from metadata.ingestion.source.database.query_parser_source import QueryParserSource
+from metadata.ingestion.source.models import TableView
+from metadata.utils import fqn
+from metadata.utils.db_utils import get_view_lineage
 from metadata.utils.logger import ingestion_logger
 
 logger = ingestion_logger()
 
 
-CHUNK_SIZE = 200
+CHUNK_SIZE = 100
 
-THREAD_TIMEOUT = 10 * 60
-PROCESS_TIMEOUT = 10 * 60
-# Maximum number of processes to use for parallel processing
-MAX_PROCESSES = min(multiprocessing.cpu_count(), 8)  # Limit to 8 or available CPUs
+THREAD_TIMEOUT = 3 * 60 * 10  # 30 minutes in seconds
 
 
 class LineageSource(QueryParserSource, ABC):
@@ -118,100 +120,94 @@ class LineageSource(QueryParserSource, ABC):
             yield from self.yield_table_query()
 
     @staticmethod
-    def generate_lineage_with_processes(
+    def generate_lineage_in_thread(
         producer_fn: Callable[[], Iterable[Any]],
         processor_fn: Callable[[Any, Queue], None],
-        args: Tuple[Any, ...],
         chunk_size: int = CHUNK_SIZE,
-        processor_timeout: int = PROCESS_TIMEOUT,
+        thread_timeout: int = THREAD_TIMEOUT,
+        max_threads: int = 10,  # Default maximum number of concurrent threads
     ):
         """
-        Process data in separate processes with timeout control.
+        Process data in separate daemon threads with timeout control.
 
         Args:
             producer_fn: Function that yields data chunks
             processor_fn: Function that processes data and adds results to the queue
             chunk_size: Size of chunks to process
-            processor_timeout: Maximum time in seconds to wait for a processor process
+            thread_timeout: Maximum time in seconds to wait for a processor thread
+            max_threads: Maximum number of concurrent threads to run
         """
-
-        def chunk_generator():
-            """Group items from producer into chunks of specified size."""
-            temp_chunk = []
-            for item in producer_fn():
-                temp_chunk.append(item)
-                if len(temp_chunk) >= chunk_size:
-                    yield temp_chunk
-                    temp_chunk = []
-            if temp_chunk:
-                yield temp_chunk
-
-        # Use multiprocessing Queue instead of threading Queue
         queue = Queue()
-        active_processes = []
+        active_threads = []
 
-        # Dictionary to track process start times
-        process_start_times = {}
-
-        # Start processing each chunk in a separate process
-        for _, chunk in enumerate(chunk_generator()):
-            # Start a process for processing
-            process = Process(
-                target=_process_chunk_in_subprocess,
-                args=(chunk, processor_fn, queue, *args),
-            )
-            process.daemon = True
-            process_start_times[
-                process.name
-            ] = time.time()  # Track when the process started
-            logger.info(
-                f"Process {process.name} started at {process_start_times[process.name]}"
-            )
-            active_processes.append(process)
-            process.start()
-
-        # Process results from the queue and check for timed-out processes
-        while active_processes or not queue.empty():
-            # Process any available results
+        def process_chunk(chunk):
+            """Process a chunk of data in a thread."""
             try:
-                while not queue.empty():
-                    yield queue.get_nowait()
-            except Exception as exc:
-                logger.warning(f"Error processing queue: {exc}")
+                processor_fn(chunk, queue)
+            except Exception as e:
+                logger.error(f"Error processing chunk: {e}")
                 logger.debug(traceback.format_exc())
 
-            # Check for completed or timed-out processes
-            still_active = []
-            for process in active_processes:
-                if process.is_alive():
-                    # Check if the process has timed out
-                    if (
-                        time.time() - process_start_times[process.name]
-                        > processor_timeout
-                    ):
-                        logger.warning(
-                            f"Process {process.name} timed out after {processor_timeout}s"
-                        )
-                        process.terminate()  # Force terminate the timed out process
-                    else:
-                        still_active.append(process)
-                else:
-                    # Clean up completed process
-                    process.join()
+        # Create an iterator for the chunks but don't consume it all at once
+        chunk_iterator = iter(chunk_generator(producer_fn, chunk_size))
 
-            active_processes = still_active
+        # Process results from the queue and check for timed-out threads
+        chunk_processed = False  # Flag to track if all chunks have been processed
+        ignored_threads = 0
+
+        while True:
+            # Start new threads until we reach the max_threads limit
+            while (
+                len(active_threads) + ignored_threads
+            ) < max_threads and not chunk_processed:
+                try:
+                    # Only fetch a new chunk when we're ready to create a thread
+                    chunk = next(chunk_iterator)
+                    thread = threading.Thread(target=process_chunk, args=(chunk,))
+                    thread.start_time = time.time()  # Track when the thread started
+                    thread.daemon = True
+                    active_threads.append(thread)
+                    thread.start()
+                except StopIteration:
+                    # No more chunks to process
+                    chunk_processed = True
+                    break
+
+            if ignored_threads == max_threads:
+                logger.warning(f"Max threads reached, skipping remaining threads")
+                break
+
+            # Process any available results
+            if queue.has_tasks():
+                yield from queue.process()
+
+            # Check for completed or timed-out threads
+            still_active = []
+            for thread in active_threads:
+                if thread.is_alive():
+                    # Check if the thread has timed out
+                    if time.time() - thread.start_time > thread_timeout:
+                        logger.warning(
+                            f"Thread {thread.name} timed out after {thread_timeout}s"
+                        )
+                        ignored_threads += 1
+                    else:
+                        still_active.append(thread)
+                # If thread is not alive, it has completed normally
+
+            active_threads = still_active
+
+            # Exit conditions: no more active threads and no more chunks to process
+            if not active_threads and chunk_processed:
+                break
 
             # Small pause to prevent CPU spinning
-            if active_processes:
+            if active_threads:
                 time.sleep(0.1)
 
         # Final check for any remaining results
-        try:
-            while not queue.empty():
-                yield queue.get_nowait()
-        except Exception as exc:
-            logger.warning(f"Error processing queue: {exc}")
-            logger.debug(traceback.format_exc())
+        while queue.has_tasks():
+            yield from queue.process()
 
     def yield_table_query(self) -> Iterator[TableQuery]:
         """
@@ -243,6 +239,57 @@ class LineageSource(QueryParserSource, ABC):
                             f"Error processing query_dict {query_dict}: {exc}"
                         )
 
+    def _query_already_processed(self, table_query: TableQuery) -> bool:
+        """
+        Check if a query has already been processed by validating if exists
+        in ES with lineageProcessed as True
+        """
+        checksums = self.metadata.es_get_queries_with_lineage(
+            service_name=table_query.serviceName,
+        )
+        return fqn.get_query_checksum(table_query.query) in checksums or {}
+
+    def query_lineage_generator(
+        self, table_queries: List[TableQuery], queue: Queue
+    ) -> Iterable[Either[Union[AddLineageRequest, CreateQueryRequest]]]:
+        if self.graph is None and self.source_config.enableTempTableLineage:
+            import networkx as nx
+
+            # Create a directed graph
+            self.graph = nx.DiGraph()
+
+        for table_query in table_queries or []:
+            if not self._query_already_processed(table_query):
+                lineages: Iterable[Either[AddLineageRequest]] = get_lineage_by_query(
+                    self.metadata,
+                    query=table_query.query,
+                    service_name=table_query.serviceName,
+                    database_name=table_query.databaseName,
+                    schema_name=table_query.databaseSchema,
+                    dialect=self.dialect,
+                    timeout_seconds=self.source_config.parsingTimeoutLimit,
+                    graph=self.graph,
+                )
+
+                for lineage_request in lineages or []:
+                    queue.put(lineage_request)
+
+                    # If we identified lineage properly, ingest the original query
+                    if lineage_request.right:
+                        queue.put(
+                            Either(
+                                right=CreateQueryRequest(
+                                    query=SqlQuery(table_query.query),
+                                    query_type=table_query.query_type,
+                                    duration=table_query.duration,
+                                    processedLineage=True,
+                                    service=FullyQualifiedEntityName(
+                                        self.config.serviceName
+                                    ),
+                                )
+                            )
+                        )
+
     def yield_query_lineage(
         self,
     ) -> Iterable[Either[Union[AddLineageRequest, CreateQueryRequest]]]:
@@ -254,19 +301,50 @@ class LineageSource(QueryParserSource, ABC):
         connection_type = str(self.service_connection.type.value)
         self.dialect = ConnectionTypeDialectMapper.dialect_of(connection_type)
         producer_fn = self.get_table_query
-        processor_fn = query_lineage_generator
-        args = (
-            self.metadata,
-            self.dialect,
-            self.graph,
-            self.source_config.parsingTimeoutLimit,
-            self.config.serviceName,
-        )
-        yield from self.generate_lineage_with_processes(
+        processor_fn = self.query_lineage_generator
+        yield from self.generate_lineage_in_thread(
             producer_fn,
             processor_fn,
-            args,
+            max_threads=self.source_config.threads,
         )
+
+    def view_lineage_generator(
+        self, views: List[TableView], queue: Queue
+    ) -> Iterable[Either[AddLineageRequest]]:
+        try:
+            for view in views:
+                for lineage in get_view_lineage(
+                    view=view,
+                    metadata=self.metadata,
+                    service_name=self.config.serviceName,
+                    connection_type=self.service_connection.type.value,
+                    timeout_seconds=self.source_config.parsingTimeoutLimit,
+                ):
+                    if lineage.right is not None:
+                        view_fqn = fqn.build(
+                            metadata=self.metadata,
+                            entity_type=Table,
+                            service_name=self.service_name,
+                            database_name=view.db_name,
+                            schema_name=view.schema_name,
+                            table_name=view.table_name,
+                            skip_es_search=True,
+                        )
+                        queue.put(
+                            Either(
+                                right=OMetaLineageRequest(
+                                    lineage_request=lineage.right,
+                                    override_lineage=self.source_config.overrideViewLineage,
+                                    entity_fqn=view_fqn,
+                                    entity=Table,
+                                )
+                            )
+                        )
+                    else:
+                        queue.put(lineage)
+        except Exception as exc:
+            logger.debug(traceback.format_exc())
+            logger.warning(f"Error processing view {view}: {exc}")
 
     def yield_view_lineage(self) -> Iterable[Either[AddLineageRequest]]:
         logger.info("Processing View Lineage")
@@ -275,21 +353,16 @@ class LineageSource(QueryParserSource, ABC):
             self.config.serviceName,
             self.source_config.incrementalLineageProcessing,
         )
-        processor_fn = view_lineage_generator
-        args = (
-            self.metadata,
-            self.config.serviceName,
-            self.service_connection.type.value,
-            self.source_config.parsingTimeoutLimit,
-            self.source_config.overrideViewLineage,
+        processor_fn = self.view_lineage_generator
+        yield from self.generate_lineage_in_thread(
+            producer_fn, processor_fn, max_threads=self.source_config.threads
         )
-        yield from self.generate_lineage_with_processes(producer_fn, processor_fn, args)
 
     def yield_procedure_lineage(
         self,
     ) -> Iterable[Either[Union[AddLineageRequest, CreateQueryRequest]]]:
         """
-        By default stored   procedure lineage is not supported.
+        By default stored procedure lineage is not supported.
         """
         logger.info(
             f"Processing Procedure Lineage not supported for {str(self.service_connection.type.value)}"
@@ -316,7 +389,6 @@ class LineageSource(QueryParserSource, ABC):
         except Exception as exc:
             logger.debug(f"Error to get column lineage: {exc}")
             logger.debug(traceback.format_exc())
-        return []
 
     def get_add_cross_database_lineage_request(
         self,
@@ -324,9 +396,6 @@ class LineageSource(QueryParserSource, ABC):
         to_entity: Table,
         column_lineage: List[ColumnLineage] = None,
     ) -> Optional[Either[AddLineageRequest]]:
-        """
-        Get the add cross database lineage request
-        """
         if from_entity and to_entity:
             return Either(
                 right=AddLineageRequest(
@@ -362,9 +431,6 @@ class LineageSource(QueryParserSource, ABC):
         Based on the query logs, prepare the lineage
         and send it to the sink
         """
-        if self.graph is None and self.source_config.enableTempTableLineage:
-            # Create a directed graph
-            self.graph = nx.DiGraph()
         if self.source_config.processViewLineage:
             yield from self.yield_view_lineage() or []
         if self.source_config.processStoredProcedureLineage:
@@ -384,3 +450,18 @@ class LineageSource(QueryParserSource, ABC):
             and self.source_config.crossDatabaseServiceNames
         ):
             yield from self.yield_cross_database_lineage() or []
+
+
+def chunk_generator(producer_fn, chunk_size):
+    """
+    Group items from producer into chunks of specified size.
+    This is a separate function to allow for better lazy evaluation.
+    """
+    temp_chunk = []
+    for item in producer_fn():
+        temp_chunk.append(item)
+        if len(temp_chunk) >= chunk_size:
+            yield temp_chunk
+            temp_chunk = []
+    if temp_chunk:
+        yield temp_chunk
