+++ ingestion/src/metadata/utils/db_utils.py
@@ -12,7 +12,7 @@
 """
 Helpers module for db sources
 """
-import time
+
 import traceback
 from typing import Iterable
 
@@ -32,7 +32,6 @@ from metadata.ingestion.lineage.sql_lineage import (
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
 from metadata.ingestion.source.models import TableView
 from metadata.utils import fqn
-from metadata.utils.execution_time_tracker import calculate_execution_time_generator
 from metadata.utils.logger import utils_logger
 
 logger = utils_logger()
@@ -48,7 +47,6 @@ def get_host_from_host_port(uri: str) -> str:
     return uri.split(":")[0]
 
 
-@calculate_execution_time_generator()
 def get_view_lineage(
     view: TableView,
     metadata: OpenMetadata,
@@ -83,8 +81,6 @@ def get_view_lineage(
     try:
         connection_type = str(connection_type)
         dialect = ConnectionTypeDialectMapper.dialect_of(connection_type)
-        start_time = time.time()
-        logger.debug(f"Processing view lineage for: {table_fqn}")
         lineage_parser = LineageParser(
             view_definition, dialect, timeout_seconds=timeout_seconds
         )
@@ -93,10 +89,6 @@ def get_view_lineage(
             # For Postgres, if schema is not defined, we need to use the public schema
             schema_name = PUBLIC_SCHEMA
 
-        end_time = time.time()
-        logger.debug(
-            f"Time taken to parse view lineage for: {table_fqn} is {end_time - start_time} seconds"
-        )
         if lineage_parser.source_tables and lineage_parser.target_tables:
             yield from get_lineage_by_query(
                 metadata,
@@ -107,7 +99,6 @@ def get_view_lineage(
                 dialect=dialect,
                 timeout_seconds=timeout_seconds,
                 lineage_source=LineageSource.ViewLineage,
-                lineage_parser=lineage_parser,
             ) or []
 
         else:
@@ -121,7 +112,6 @@ def get_view_lineage(
                 dialect=dialect,
                 timeout_seconds=timeout_seconds,
                 lineage_source=LineageSource.ViewLineage,
-                lineage_parser=lineage_parser,
             ) or []
     except Exception as exc:
         logger.debug(traceback.format_exc())
